% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{addClassAttribute}
\alias{addClassAttribute}
\alias{addClassAttribute,ANY,character-method}
\title{Add Class Attribute}
\usage{
addClassAttribute(obj, class_name)

\S4method{addClassAttribute}{ANY,character}(obj, class_name)
}
\arguments{
\item{obj}{\strong{Signature argument}.
Object to be updated.}

\item{class_name}{\strong{Signature argument}.
Object containing class name information.}

\item{obj}{\code{\link{ANY}}.}

\item{class_name}{\code{\link{character}}.}
}
\value{
Object with updated class graph (\code{c(<class_name>, <class>)}).
}
\description{
Adds class attribute to an object and returns it.
}
\section{Methods (by class)}{
\itemize{
\item \code{obj = ANY,class_name = character}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.obj}
}

