% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{asExpandedRepository}
\alias{asExpandedRepository}
\alias{asExpandedRepository,RappExpandedPackageRepositoryS3-method}
\alias{asExpandedRepository,RappPackageRepositoryGenericS3-method}
\alias{asExpandedRepository,RappPackageRepositoryMacBinaryS3-method}
\alias{asExpandedRepository,RappPackageRepositoryS3-method}
\alias{asExpandedRepository,RappPackageRepositorySourceS3-method}
\alias{asExpandedRepository,RappPackageRepositoryWinBinaryS3-method}
\alias{asExpandedRepository,character-method}
\alias{asExpandedRepository,missing-method}
\title{As Expanded Repository}
\usage{
asExpandedRepository(repos = asRepository("."), ensure = FALSE, ...)


  \S4method{asExpandedRepository}{RappPackageRepositoryS3}(repos = asRepository("."),
  ensure = FALSE, ...)

\S4method{asExpandedRepository}{missing}(repos = asRepository("."),
  ensure = FALSE, ...)

\S4method{asExpandedRepository}{character}(repos = asRepository("."),
  ensure = FALSE, ...)


  \S4method{asExpandedRepository}{RappExpandedPackageRepositoryS3}(repos = asRepository("."),
  ensure = FALSE, ...)


  \S4method{asExpandedRepository}{RappPackageRepositoryGenericS3}(repos = asRepository("."),
  ensure = FALSE, ...)


  \S4method{asExpandedRepository}{RappPackageRepositoryMacBinaryS3}(repos = asRepository("."),
  ensure = FALSE, ...)


  \S4method{asExpandedRepository}{RappPackageRepositoryWinBinaryS3}(repos = asRepository("."),
  ensure = FALSE, ...)


  \S4method{asExpandedRepository}{RappPackageRepositorySourceS3}(repos = asRepository("."),
  ensure = FALSE, ...)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repos information.}

\item{ensure}{\code{logical}.
Ensure repository existence (\code{TRUE}) or not (\code{FALSE}, default).}

\item{...}{Further arguments passed to:
\code{\link[rapp.core.repos]{getPartialRepositoryScaffold}},
\code{\link[rapp.core.repos]{ensureRepository}}..}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{missing}}.}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
\code{RappExpandedPackageRepositoryS3}. Object containing all
   repository sublevels with respective class attributes
   (\code{RappPackageRepositoryMacBinaryS3},
   \code{RappPackageRepositoryWinBinaryS3},
   \code{RappPackageRepositorySourceS3}.

TODO

TODO

TODO

TODO

TODO

TODO
}
\description{
Expands actual repository paths and returns an object of class
\code{\link[rapp.core.repos]{RappExpandedRepositoryS3}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{RappPackageRepositoryS3}: 

\item \code{missing}: 

\item \code{character}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{
  
## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

