% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{asRepository}
\alias{asRepository}
\alias{asRepository,RappExpandedPackageRepositoryS3-method}
\alias{asRepository,RappPackageRepositoryGenericS3-method}
\alias{asRepository,RappPackageRepositoryMacBinaryS3-method}
\alias{asRepository,RappPackageRepositoryS3-method}
\alias{asRepository,RappPackageRepositorySourceS3-method}
\alias{asRepository,RappPackageRepositoryWinBinaryS3-method}
\alias{asRepository,character-method}
\alias{asRepository,missing-method}
\title{As Repository}
\usage{
asRepository(repos = ".", ensure = FALSE, ...)

\S4method{asRepository}{character}(repos = ".", ensure = FALSE, ...)

\S4method{asRepository}{missing}(repos = ".", ensure = FALSE, ...)

\S4method{asRepository}{RappPackageRepositoryS3}(repos = ".",
  ensure = FALSE, ...)

\S4method{asRepository}{RappExpandedPackageRepositoryS3}(repos = ".",
  ensure = FALSE, ...)

\S4method{asRepository}{RappPackageRepositoryGenericS3}(repos = ".",
  ensure = FALSE, ...)

\S4method{asRepository}{RappPackageRepositoryMacBinaryS3}(repos = ".",
  ensure = FALSE, ...)

\S4method{asRepository}{RappPackageRepositoryWinBinaryS3}(repos = ".",
  ensure = FALSE, ...)

\S4method{asRepository}{RappPackageRepositorySourceS3}(repos = ".",
  ensure = FALSE, ...)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repos information.}

\item{ensure}{\code{logical}.
Ensure repository existence (\code{TRUE}) or not (\code{FALSE}, default).}

\item{...}{Further arguments passed to:
\code{\link[rapp.core.repos]{ensureRepository}}.}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{missing}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
\code{RappPackageRepositoryS3}. Identical to \code{repos} with updated
		class table.
}
\description{
Turns a \code{character} path of a package repository into an object
of class \code{RappPackageRepositoryS3}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

