% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{buildIntoRepositoryInfrastructure}
\alias{buildIntoRepositoryInfrastructure}
\alias{buildIntoRepositoryInfrastructure,character-method}
\title{Build Into Repository Infrastructure}
\usage{
buildIntoRepositoryInfrastructure(repos_home = file.path(Sys.getenv("HOME"),
  ".rapp/repos"), binary = FALSE, remove_old = FALSE, ...)


  \S4method{buildIntoRepositoryInfrastructure}{character}(repos_home = file.path(Sys.getenv("HOME"),
  ".rapp/repos"), binary = FALSE, remove_old = FALSE, ...)
}
\arguments{
\item{repos_home}{\strong{Signature argument}.
Object containing information about the repository parent location.}

\item{binary}{\code{\link{logical}}.
In addition to the source version, also build binary version
(\code{TRUE}) or not (\code{FALSE} (default)).}

\item{remove_old}{\code{\link{logical}}.
Remove old content (\code{TRUE}) or not (\code{FALSE} (default)).}

\item{repos_home}{\code{\link{character}}.}
}
\value{
TODO
}
\description{
Build into repository infrastructure as ensured by
\code{\link[rapp.core.repos]{ensureRepositoryInfrastructure}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

