% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ensureRepositoryIndex}
\alias{ensureRepositoryIndex}
\alias{ensureRepositoryIndex,RappExpandedPackageRepositoryS3-method}
\alias{ensureRepositoryIndex,RappPackageRepositoryGenericS3-method}
\alias{ensureRepositoryIndex,RappPackageRepositoryMacBinaryS3-method}
\alias{ensureRepositoryIndex,RappPackageRepositoryS3-method}
\alias{ensureRepositoryIndex,RappPackageRepositorySourceS3-method}
\alias{ensureRepositoryIndex,RappPackageRepositoryWinBinaryS3-method}
\alias{ensureRepositoryIndex,character-method}
\title{Ensure Repository Registry Files}
\usage{
ensureRepositoryIndex(repos, ...)

\S4method{ensureRepositoryIndex}{character}(repos, ...)

\S4method{ensureRepositoryIndex}{RappPackageRepositoryS3}(repos, ...)

\S4method{ensureRepositoryIndex}{RappExpandedPackageRepositoryS3}(repos, ...)

\S4method{ensureRepositoryIndex}{RappPackageRepositoryGenericS3}(repos,
  type = c("mac.binary", "source", "win.binary"), ...)

\S4method{ensureRepositoryIndex}{RappPackageRepositoryMacBinaryS3}(repos, ...)

\S4method{ensureRepositoryIndex}{RappPackageRepositorySourceS3}(repos, ...)

\S4method{ensureRepositoryIndex}{RappPackageRepositoryWinBinaryS3}(repos, ...)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}
}
\value{
TODO

TODO

TODO
}
\description{
Ensures existence of a repository's registry files (\code{PACKAGES} and
\code{PACKAGES.gz}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}
\seealso{
\code{\link[rapp2]{ensureRepositoryIndex-character-method}}
}

