% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ensureRepositoryRegistryFiles}
\alias{ensureRepositoryRegistryFiles}
\alias{ensureRepositoryRegistryFiles,character-method}
\title{Ensure Repository Registry Files}
\usage{
ensureRepositoryRegistryFiles(repos)

\S4method{ensureRepositoryRegistryFiles}{character}(repos)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{repos}{\code{\link{character}}.}
}
\value{
TODO
}
\description{
Ensures existence of a repository's registry files (\code{PACKAGES} and
\code{PACKAGES.gz}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}
\seealso{
\code{\link[rapp2]{ensureRepositoryRegistryFiles-character-method}}
}

