% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getExpandedRepository}
\alias{getExpandedRepository}
\alias{getExpandedRepository,character-method}
\alias{getExpandedRepository,missing-method}
\title{Get Repository Scaffold}
\usage{
getExpandedRepository(path = ".", ...)

\S4method{getExpandedRepository}{character}(path = ".", ...)

\S4method{getExpandedRepository}{missing}(path = ".", ...)
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{...}{Further arguments passed to:
\code{\link{getPartialRepositoryScaffold}}.}

\item{path}{\code{\link{character}}.}

\item{path}{\code{\link{missing}}.}
}
\value{
TODO

TODO
}
\description{
Provides scaffold of a package repository as expected by R.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}
\seealso{
\code{\link[rapp2]{getExpandedRepository-character-method}}
}

