% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getPackageBuildFilePatterns}
\alias{getPackageBuildFilePatterns}
\alias{getPackageBuildFilePatterns,RappExpandedPackageRepositoryS3-method}
\alias{getPackageBuildFilePatterns,RappPackageRepositoryGenericS3-method}
\alias{getPackageBuildFilePatterns,RappPackageRepositoryMacBinaryS3-method}
\alias{getPackageBuildFilePatterns,RappPackageRepositoryS3-method}
\alias{getPackageBuildFilePatterns,RappPackageRepositorySourceS3-method}
\alias{getPackageBuildFilePatterns,RappPackageRepositoryWinBinaryS3-method}
\alias{getPackageBuildFilePatterns,RappParsedPackageRepositoryIndexS3-method}
\alias{getPackageBuildFilePatterns,character-method}
\alias{getPackageBuildFilePatterns,missing-method}
\title{Get Package Versions based on Repository Index File}
\usage{
getPackageBuildFilePatterns(input = "PACKAGES", ...)

\S4method{getPackageBuildFilePatterns}{missing}(input = "PACKAGES", ...)

\S4method{getPackageBuildFilePatterns}{character}(input = "PACKAGES", ...)


  \S4method{getPackageBuildFilePatterns}{RappParsedPackageRepositoryIndexS3}(input = "PACKAGES",
  ...)


  \S4method{getPackageBuildFilePatterns}{RappPackageRepositoryS3}(input = "PACKAGES",
  ...)


  \S4method{getPackageBuildFilePatterns}{RappExpandedPackageRepositoryS3}(input = "PACKAGES",
  ...)


  \S4method{getPackageBuildFilePatterns}{RappPackageRepositoryGenericS3}(input = "PACKAGES",
  ...)


  \S4method{getPackageBuildFilePatterns}{RappPackageRepositoryMacBinaryS3}(input = "PACKAGES",
  ...)

\S4method{getPackageBuildFilePatterns}{RappPackageRepositoryWinBinaryS3}(input,
  type, ...)

\S4method{getPackageBuildFilePatterns}{RappPackageRepositorySourceS3}(input,
  type, ...)
}
\arguments{
\item{input}{\strong{Signature argument}.
Object containing source input for parsing.}

\item{...}{Further arguments passed to:
\code{\link[repositr]{asExpandedRepository}},
\code{\link[repositr]{hasRepositoryPackages}}.}

\item{input}{\code{\link{missing}}.}

\item{input}{\code{\link{character}}.}

\item{input}{\code{\link{RappParsedPackageRepositoryIndexS3}}.}

\item{input}{\code{\link{RappPackageRepositoryS3}}.}

\item{input}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{input}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{input}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{input}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{input}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
TODO

TODO

TODO

TODO
}
\description{
Get package versions based on a repository's index file (\code{PACKAGES}).
}
\section{Methods (by class)}{
\itemize{
\item \code{missing}: 

\item \code{character}: 

\item \code{RappParsedPackageRepositoryIndexS3}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/repositr}
}

