% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getPartialRepositoryScaffold}
\alias{getPartialRepositoryScaffold}
\alias{getPartialRepositoryScaffold,character-method}
\alias{getPartialRepositoryScaffold,missing-method}
\title{Get Partial Repository Scaffold}
\usage{
getPartialRepositoryScaffold(rversion = paste(R.version$major,
  unlist(strsplit(R.version$minor, split = "\\\\."))[2], sep = "."))


  \S4method{getPartialRepositoryScaffold}{character}(rversion = paste(R.version$major,
  unlist(strsplit(R.version$minor, split = "\\\\."))[2], sep = "."))


  \S4method{getPartialRepositoryScaffold}{missing}(rversion = paste(R.version$major,
  unlist(strsplit(R.version$minor, split = "\\\\."))[2], sep = "."))
}
\arguments{
\item{rversion}{\strong{Signature argument}.
Object containing R version information.}

\item{rversion}{\code{\link{character}}.}

\item{rversion}{\code{\link{missing}}.}
}
\value{
TODO

TODO
}
\description{
Provides partial scaffold of a package repository as expected by R.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.packrat}
}
\seealso{
\code{\link[rapp2]{getPartialRepositoryScaffold-character-method}}
}

