% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getRepositoryInfrastructurePaths}
\alias{getRepositoryInfrastructurePaths}
\alias{getRepositoryInfrastructurePaths,character,character,character-method}
\alias{getRepositoryInfrastructurePaths,character,missing,missing-method}
\title{Get Repository Infrastructure Paths}
\usage{
getRepositoryInfrastructurePaths(repos_home = file.path(Sys.getenv("HOME"),
  ".rapp/repos"), pkg_name = character(), pkg_version = character(), ...)


  \S4method{getRepositoryInfrastructurePaths}{character,missing,missing}(repos_home = file.path(Sys.getenv("HOME"),
  ".rapp/repos"), pkg_name = character(), pkg_version = character(), ...)


  \S4method{getRepositoryInfrastructurePaths}{character,character,character}(repos_home = file.path(Sys.getenv("HOME"),
  ".rapp/repos"), pkg_name = character(), pkg_version = character(), ...)
}
\arguments{
\item{repos_home}{\strong{Signature argument}.
Object containing information about the repository parent location.}

\item{pkg_name}{\strong{Signature argument}.
Object containing package name information.}

\item{pkg_version}{\strong{Signature argument}.
Object containing package version information.}

\item{repos_home}{\code{\link{character}}.}

\item{pkg_name}{\code{\link{missing}}.}

\item{pkg_version}{\code{\link{missing}}.}

\item{repos_home}{\code{\link{character}}.}

\item{pkg_name}{\code{\link{character}}.}

\item{pkg_version}{\code{\link{character}}.}
}
\value{
TODO

TODO
}
\description{
Get repository infrastructure paths.
}
\section{Methods (by class)}{
\itemize{
\item \code{repos_home = character,pkg_name = missing,pkg_version = missing}: 

\item \code{repos_home = character,pkg_name = character,pkg_version = character}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

