% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getRepositoryPathByType}
\alias{getRepositoryPathByType}
\alias{getRepositoryPathByType,character,character-method}
\alias{getRepositoryPathByType,character,missing-method}
\title{Select Repository Path By Type}
\usage{
getRepositoryPathByType(repos, type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{character,character}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{character,missing}(repos,
  type = c("win.binary", "source", "mac.binary"))
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{type}{\strong{Signature argument}.
Object containing repository type information.}

\item{repos}{\code{\link{character}}.}

\item{type}{\code{\link{character}}.}

\item{repos}{\code{\link{character}}.}

\item{type}{\code{\link{missing}}.}
}
\value{
\code{\link{character}}. Repository path

\code{\link{character}}. Repository path
}
\description{
Selects repository path based on tyype.
}
\section{Methods (by class)}{
\itemize{
\item \code{repos = character,type = character}: 

\item \code{repos = character,type = missing}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

