% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getRepositoryRoot}
\alias{getRepositoryRoot}
\alias{getRepositoryRoot,RappExpandedPackageRepositoryS3-method}
\alias{getRepositoryRoot,RappPackageRepositoryGenericS3-method}
\alias{getRepositoryRoot,RappPackageRepositoryMacBinaryS3-method}
\alias{getRepositoryRoot,RappPackageRepositoryS3-method}
\alias{getRepositoryRoot,RappPackageRepositorySourceS3-method}
\alias{getRepositoryRoot,RappPackageRepositoryWinBinaryS3-method}
\alias{getRepositoryRoot,character-method}
\alias{getRepositoryRoot,missing-method}
\title{Get Repository's Root Path}
\usage{
getRepositoryRoot(repos = ".")

\S4method{getRepositoryRoot}{character}(repos = ".")

\S4method{getRepositoryRoot}{missing}(repos = ".")

\S4method{getRepositoryRoot}{RappPackageRepositoryS3}(repos = ".")

\S4method{getRepositoryRoot}{RappExpandedPackageRepositoryS3}(repos = ".")

\S4method{getRepositoryRoot}{RappPackageRepositoryGenericS3}(repos = ".")

\S4method{getRepositoryRoot}{RappPackageRepositoryMacBinaryS3}(repos = ".")

\S4method{getRepositoryRoot}{RappPackageRepositoryWinBinaryS3}(repos = ".")

\S4method{getRepositoryRoot}{RappPackageRepositorySourceS3}(repos = ".")
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repos information.}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{missing}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
\code{RappPackageRepositoryS3}. No matter what type repository has
   been passed in, the method returns only the root part of the repository.
}
\description{
Get a repository's root path.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

