% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getRepositoryType}
\alias{getRepositoryType}
\alias{getRepositoryType,character-method}
\title{Get Repository Type}
\usage{
getRepositoryType(repos)

\S4method{getRepositoryType}{character}(repos)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{repos}{\code{\link{character}}.}
}
\value{
TODO
}
\description{
Infers the repository type based on repository location.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

