% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{hasRepositoryPackages}
\alias{hasRepositoryPackages}
\alias{hasRepositoryPackages,RappExpandedPackageRepositoryS3-method}
\alias{hasRepositoryPackages,RappPackageRepositoryGenericS3-method}
\alias{hasRepositoryPackages,RappPackageRepositoryMacBinaryS3-method}
\alias{hasRepositoryPackages,RappPackageRepositoryS3-method}
\alias{hasRepositoryPackages,RappPackageRepositorySourceS3-method}
\alias{hasRepositoryPackages,RappPackageRepositoryWinBinaryS3-method}
\alias{hasRepositoryPackages,character-method}
\alias{hasRepositoryPackages,missing-method}
\title{Exist Packages In Repository}
\usage{
hasRepositoryPackages(repos = ".", ...)

\S4method{hasRepositoryPackages}{missing}(repos = ".", ...)

\S4method{hasRepositoryPackages}{character}(repos = ".", ...)

\S4method{hasRepositoryPackages}{RappPackageRepositoryS3}(repos = ".", ...)


  \S4method{hasRepositoryPackages}{RappExpandedPackageRepositoryS3}(repos = ".",
  ...)

\S4method{hasRepositoryPackages}{RappPackageRepositoryGenericS3}(repos = ".",
  ...)


  \S4method{hasRepositoryPackages}{RappPackageRepositoryMacBinaryS3}(repos = ".",
  ...)


  \S4method{hasRepositoryPackages}{RappPackageRepositoryWinBinaryS3}(repos = ".",
  ...)

\S4method{hasRepositoryPackages}{RappPackageRepositorySourceS3}(repos = ".",
  ...)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{further}{arguments passed to:
\code{\link[repositr]{asRepository}},
\code{\link[repositr]{asExpandedRepository}}.}

\item{repos}{\code{\link{missing}}.}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
\code{logical}. For each package type, \code{TRUE} if any packages
   exist in the respective sub repository,\code{FALSE} if not.

\code{logical}. For each package type, \code{TRUE} if any packages
   exist in the respective sub repository,\code{FALSE} if not.

\code{logical}. For each package type, \code{TRUE} if any packages
   exist in the respective sub repository,\code{FALSE} if not.

\code{logical}. For each package type, \code{TRUE} if any packages
   exist in the respective sub repository,\code{FALSE} if not.

\code{logical}. For each package type, \code{TRUE} if any packages
   exist in the respective sub repository,\code{FALSE} if not.
}
\description{
Checks if a repository, or to be more precise, the actual sublevel repositories
(for binary and source builds), contain any packages at all by investigating
its index files (\code{PACKAGES}).
}
\section{Methods (by class)}{
\itemize{
\item \code{missing}: 

\item \code{character}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/repositr}
}

