% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{normalizeRepositoryPath}
\alias{normalizeRepositoryPath}
\alias{normalizeRepositoryPath,RappPackageRepositoryS3,character-method}
\alias{normalizeRepositoryPath,RappPackageRepositoryS3,missing-method}
\alias{normalizeRepositoryPath,character,character-method}
\alias{normalizeRepositoryPath,character,missing-method}
\title{Normalize Repository Path}
\usage{
normalizeRepositoryPath(repos = ".", type = c("fs", "url_file", "url_http",
  "url_ftp"))

\S4method{normalizeRepositoryPath}{character,character}(repos = ".",
  type = c("fs", "url_file", "url_http", "url_ftp"))

\S4method{normalizeRepositoryPath}{character,missing}(repos = ".",
  type = c("fs", "url_file", "url_http", "url_ftp"))


  \S4method{normalizeRepositoryPath}{RappPackageRepositoryS3,character}(repos = ".",
  type = c("fs", "url_file", "url_http", "url_ftp"))


  \S4method{normalizeRepositoryPath}{RappPackageRepositoryS3,missing}(repos = ".",
  type = c("fs", "url_file", "url_http", "url_ftp"))
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository path information.}

\item{type}{\strong{Signature argument}.
Object containing normalization type information.}

\item{repos}{\code{\link{character}}.}

\item{type}{\code{\link{character}}.}

\item{repos}{\code{\link{character}}.}

\item{type}{\code{\link{missing}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{type}{\code{\link{character}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{type}{\code{\link{missing}}.}
}
\value{
\code{\link{character}}. Normalized path.

\code{\link{character}}. Normalized path.

\code{\link{character}}. Normalized path.

\code{\link{character}}. Normalized path.
}
\description{
Normalizes repository paths.
}
\section{Methods (by class)}{
\itemize{
\item \code{repos = character,type = character}: 

\item \code{repos = character,type = missing}: 

\item \code{repos = RappPackageRepositoryS3,type = character}: 

\item \code{repos = RappPackageRepositoryS3,type = missing}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

