% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{parseRepositoryIndexFile}
\alias{parseRepositoryIndexFile}
\alias{parseRepositoryIndexFile,character-method}
\alias{parseRepositoryIndexFile,missing-method}
\title{Parse Repository Index File}
\usage{
parseRepositoryIndexFile(path = "PACKAGES")

\S4method{parseRepositoryIndexFile}{missing}(path = "PACKAGES")

\S4method{parseRepositoryIndexFile}{character}(path = "PACKAGES")
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing source information for parsing.}

\item{path}{\code{\link{missing}}.}

\item{path}{\code{\link{character}}.}
}
\description{
Parses a repository's index file (\code{PACKAGES}).
}
\section{Methods (by class)}{
\itemize{
\item \code{missing}: 

\item \code{character}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/repositr}
}

