% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{refreshRepository}
\alias{refreshRepository}
\alias{refreshRepository,character-method}
\title{Refresh Repository}
\usage{
refreshRepository(repos, remove_old = FALSE)

\S4method{refreshRepository}{character}(repos, remove_old = FALSE)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{remove_old}{\code{\link{logical}}.
Remove old content (\code{TRUE}) or not (\code{FALSE} (default)).}

\item{repos}{\code{\link{character}}.}
}
\value{
\code{\link{logical}}.
}
\description{
Refreshes repository.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

