% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{refreshRepositoryIndex}
\alias{refreshRepositoryIndex}
\alias{refreshRepositoryIndex,RappExpandedPackageRepositoryS3-method}
\alias{refreshRepositoryIndex,RappPackageRepositoryGenericS3-method}
\alias{refreshRepositoryIndex,RappPackageRepositoryMacBinaryS3-method}
\alias{refreshRepositoryIndex,RappPackageRepositoryS3-method}
\alias{refreshRepositoryIndex,RappPackageRepositorySourceS3-method}
\alias{refreshRepositoryIndex,RappPackageRepositoryWinBinaryS3-method}
\alias{refreshRepositoryIndex,character-method}
\title{Refresh Repository}
\usage{
refreshRepositoryIndex(repos, ...)

\S4method{refreshRepositoryIndex}{character}(repos, ...)

\S4method{refreshRepositoryIndex}{RappPackageRepositoryS3}(repos, ...)

\S4method{refreshRepositoryIndex}{RappExpandedPackageRepositoryS3}(repos, ...)

\S4method{refreshRepositoryIndex}{RappPackageRepositoryGenericS3}(repos, type,
  ...)

\S4method{refreshRepositoryIndex}{RappPackageRepositoryMacBinaryS3}(repos, ...)

\S4method{refreshRepositoryIndex}{RappPackageRepositoryWinBinaryS3}(repos, ...)

\S4method{refreshRepositoryIndex}{RappPackageRepositorySourceS3}(repos, ...)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
\code{\link{logical}}.

\code{\link{logical}}.

\code{\link{logical}}.

TODO

TODO

TODO

TODO
}
\description{
Refreshes repository.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

