% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{removeOldRepositoryPackages}
\alias{removeOldRepositoryPackages}
\alias{removeOldRepositoryPackages,RappExpandedPackageRepositoryS3-method}
\alias{removeOldRepositoryPackages,RappPackageRepositoryGenericS3-method}
\alias{removeOldRepositoryPackages,RappPackageRepositoryMacBinaryS3-method}
\alias{removeOldRepositoryPackages,RappPackageRepositoryS3-method}
\alias{removeOldRepositoryPackages,RappPackageRepositorySourceS3-method}
\alias{removeOldRepositoryPackages,RappPackageRepositoryWinBinaryS3-method}
\alias{removeOldRepositoryPackages,character-method}
\title{Remove Old Packages From Repository}
\usage{
removeOldRepositoryPackages(repos, type = c("mac.binary", "source",
  "win.binary"), ...)

\S4method{removeOldRepositoryPackages}{character}(repos,
  type = c("mac.binary", "source", "win.binary"), ...)

\S4method{removeOldRepositoryPackages}{RappPackageRepositoryS3}(repos,
  type = c("mac.binary", "source", "win.binary"), ...)

\S4method{removeOldRepositoryPackages}{RappExpandedPackageRepositoryS3}(repos,
  type = c("mac.binary", "source", "win.binary"), ...)

\S4method{removeOldRepositoryPackages}{RappPackageRepositoryGenericS3}(repos,
  type, pattern = character(), ...)

\S4method{removeOldRepositoryPackages}{RappPackageRepositoryMacBinaryS3}(repos,
  type = c("mac.binary", "source", "win.binary"), ...)

\S4method{removeOldRepositoryPackages}{RappPackageRepositoryWinBinaryS3}(repos,
  type = c("mac.binary", "source", "win.binary"), ...)

\S4method{removeOldRepositoryPackages}{RappPackageRepositorySourceS3}(repos,
  type = c("mac.binary", "source", "win.binary"), ...)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{type}{\code{character}.
Package types to be removed. Any subset of \code{c("mac.binary",
"source", "win.binary")}. If missing, \strong{all} package types are
considered.}

\item{...}{Further arguments passed to:
\code{\link[repositr]{asExpandedRepository}}.}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryGenericS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
\code{\link{logical}}.

\code{\link{logical}}.

\code{\link{logical}}.

TODO

TODO

TODO

TODO
}
\description{
Removes outdated package builds from a repository.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryGenericS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/repositr}
}

