% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{repositr}
\alias{repositr}
\alias{repositr-package}
\title{Local Package Repository Management}
\description{
Functionality for managing local package repositories.
The primary goal behind the development of this package was a systematic
way for managing own packages in local package repositories in order to
seamlessly integrate them with packrat
}
\details{
The core functions/methods of this package:
 \itemize{
   \item{\code{\link[repositr]{ensureRepository}}: }{

     Ensures the existence of local repositories. To be more precise, it
     creates a local repository infrastructure with a \emph{global}
     repository and version-specific repositories for each package that
     is build with \code{\link[repositr]{buildIntoRepositoryInfrastructure}}
   }
   \item{\code{\link[repositr]{buildIntoRepositoryInfrastructure}}: }{

     Builds a package and ensures that it is available in the local repository
     infrastructure.
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/repositr}
}

