% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{resetLocalPackratRepository}
\alias{resetLocalPackratRepository}
\alias{resetLocalPackratRepository,character-method}
\alias{resetLocalPackratRepository,missing-method}
\title{Reset Local Packrat Repository}
\usage{
resetLocalPackratRepository(opt_name = "local.repos")

\S4method{resetLocalPackratRepository}{character}(opt_name = "local.repos")

\S4method{resetLocalPackratRepository}{missing}(opt_name = "local.repos")
}
\arguments{
\item{opt_name}{\strong{Signature argument}.
Object containing option name information.}

\item{opt_name}{\code{\link{character}}.}

\item{opt_name}{\code{\link{missing}}.}
}
\value{
\code{logical}. \code{TRUE}.

see \code{character} method.
}
\description{
Reset local packrat package repository.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

