% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{setLocalPackratRepository}
\alias{setLocalPackratRepository}
\alias{setLocalPackratRepository,RappExpandedPackageRepositoryS3-method}
\alias{setLocalPackratRepository,RappPackageRepositoryMacBinaryS3-method}
\alias{setLocalPackratRepository,RappPackageRepositoryS3-method}
\alias{setLocalPackratRepository,RappPackageRepositorySourceS3-method}
\alias{setLocalPackratRepository,RappPackageRepositoryWinBinaryS3-method}
\alias{setLocalPackratRepository,character-method}
\title{Set Local Packrat Repository}
\usage{
setLocalPackratRepository(repos, append_existing = TRUE)

\S4method{setLocalPackratRepository}{character}(repos, append_existing = TRUE)

\S4method{setLocalPackratRepository}{RappPackageRepositoryS3}(repos,
  append_existing = TRUE)

\S4method{setLocalPackratRepository}{RappExpandedPackageRepositoryS3}(repos,
  append_existing = TRUE)

\S4method{setLocalPackratRepository}{RappPackageRepositoryMacBinaryS3}(repos,
  append_existing = TRUE)

\S4method{setLocalPackratRepository}{RappPackageRepositoryWinBinaryS3}(repos,
  append_existing = TRUE)

\S4method{setLocalPackratRepository}{RappPackageRepositorySourceS3}(repos,
  append_existing = TRUE)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository path information.}

\item{append_existing}{\code{\link{logical}}.
Append value of \code{repos} to existing values of
\code{packrat::get_opts("local.repos")} (\code{TRUE}, default) or overwrite
existing content (\code{FALSE})}

\item{repos}{\code{\link{character}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}
}
\value{
\code{logical}. \code{TRUE}.

\code{logical}. \code{TRUE}.

\code{logical}. \code{TRUE}.

\code{logical}. \code{TRUE}.

\code{logical}. \code{TRUE}.

\code{logical}. \code{TRUE}.
}
\description{
Set local packrat package repository.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{RappPackageRepositoryS3}: 

\item \code{RappExpandedPackageRepositoryS3}: 

\item \code{RappPackageRepositoryMacBinaryS3}: 

\item \code{RappPackageRepositoryWinBinaryS3}: 

\item \code{RappPackageRepositorySourceS3}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/repositr}
}

