% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{signalCondition}
\alias{signalCondition}
\alias{signalCondition,character-method}
\alias{signalCondition,missing-method}
\title{Signal Condition}
\usage{
signalCondition(condition, msg = NULL, call = sys.call(-2),
  type = c("message", "warning", "error", "condition"), signal = TRUE)

\S4method{signalCondition}{character}(condition, msg = NULL,
  call = sys.call(-2), type = c("message", "warning", "error", "condition"),
  signal = TRUE)

\S4method{signalCondition}{missing}(condition, msg = NULL,
  call = sys.call(-2), type = c("message", "warning", "error", "condition"),
  signal = TRUE)
}
\arguments{
\item{condition}{\strong{Signature argument}.}

\item{msg}{\code{\link{character}}. Dimension: any.
Condition message. Named components are transformed into the following
structure: \code{* <item>: <item-value>}, (e.g. "Item 1"="some message"
will become "* Item 1: some message" in the actual condition message.
The first element will be treated as the message's header.
Line numbers are automatically added beginning from the second line
(\code{[<number>] some message} or
\code{[<number>] * <item>: <item-value>}).}

\item{call}{\code{\link{call}}. Dimension: 1.
System call one frame up the calling stack. This will correspond to
the closure that is spanned by the function/method that calls this
method when a condition is to be signaled.
Default: \code{sys.call(-1)}.
There should be no need to change this default value.}

\item{type}{\code{\link{character}}. Dimension: 1.
Selector for the condition type.
Choices: \code{"message"}, \code{"warning"}, \code{"error"} or
\code{"condition"}.
Default: \code{"message"}.}

\item{signal}{\code{\link{logical}}. Dimension: 1.
\code{TRUE} (default) means the condition is signaled right away,
\code{FALSE} means the condition object is returned only.}

\item{condition}{\code{\link{character}}. Dimension: any.
Name(s) of the condition(es) that will be passed to the condition object's
class table.}

\item{condition}{\code{\link{missing}}.}
}
\value{
Object of the class(es) that have been provided by \code{condition} and
		also inheriting from \code{\link{condition}} and depending on
		\code{type} also from \code{\link{message}}, \code{\link{warning}} or
		\code{\link{error}}.
}
\description{
Signals custom conditions.
}
\details{
The returned object will always inherit from class \code{condition}.
The value of \code{condition} will always be the first entry/entries of the
object's class table.
Depending on the setting of \code{type}, the it also inherits from
either \code{message}, \code{warning}, \code{error} or \code{custom}.
If no value for \code{condition} is specified, then depending on the setting
of \code{type} respective default values are used: \code{"DefaultMessage"},
\code{"DefaultWarning"}, \code{"DefaultError"} and \code{"DefaultCondition"}.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 

\item \code{missing}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/}
}
\seealso{
\code{\link[base]{condition}}
}

