% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{ensureRepository}
\alias{ensureRepository}
\alias{ensureRepository,character,character-method}
\alias{ensureRepository,character,missing-method}
\title{Ensure Repository}
\usage{
ensureRepository(path = ".", rversion = paste(R.version$major,
  unlist(strsplit(R.version$minor, split = "\\\\."))[2], sep = "."))

\S4method{ensureRepository}{character,character}(path = ".",
  rversion = paste(R.version$major, unlist(strsplit(R.version$minor, split =
  "\\\\."))[2], sep = "."))

\S4method{ensureRepository}{character,missing}(path = ".",
  rversion = paste(R.version$major, unlist(strsplit(R.version$minor, split =
  "\\\\."))[2], sep = "."))
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{rversion}{\strong{Signature argument}.
Object containing version information.}

\item{path}{\code{\link{character}}.}

\item{rversion}{\code{\link{character}}.}

\item{path}{\code{\link{character}}.}

\item{rversion}{\code{\link{missing}}.}
}
\value{
TODO

TODO
}
\description{
Ensures existence of an R package repository.
}
\section{Methods (by class)}{
\itemize{
\item \code{path = character,rversion = character}: 

\item \code{path = character,rversion = missing}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}
\seealso{
\code{\link[rapp2]{ensureRepository-character-method}}
}

