% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getExpandedRepositoryPaths}
\alias{getExpandedRepositoryPaths}
\alias{getExpandedRepositoryPaths,RappPackageRepositoryS3-method}
\alias{getExpandedRepositoryPaths,character-method}
\alias{getExpandedRepositoryPaths,missing-method}
\title{Get Repository Scaffold}
\usage{
getExpandedRepositoryPaths(repos = asRepository("."), type = c("fs",
  "url_file", "url_http", "url_ftp"), ...)


  \S4method{getExpandedRepositoryPaths}{RappPackageRepositoryS3}(repos = asRepository("."),
  type = c("fs", "url_file", "url_http", "url_ftp"), ...)

\S4method{getExpandedRepositoryPaths}{missing}(repos = asRepository("."),
  type = c("fs", "url_file", "url_http", "url_ftp"), ...)

\S4method{getExpandedRepositoryPaths}{character}(repos = asRepository("."),
  type = c("fs", "url_file", "url_http", "url_ftp"), ...)
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repos information.}

\item{type}{\code{\link{character}}.
Path type. One of \code{ c("fs", "url_file", "url_http", "url_ftp")}.
See \code{\link[repositr]{normalizeRepositoryPath}}.}

\item{...}{Further arguments passed to:
\code{\link[repositr]{getPartialRepositoryScaffold}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{repos}{\code{\link{missing}}.}

\item{repos}{\code{\link{character}}.}
}
\value{
\code{RappExpandedPackageRepositoryS3}. Object containing all
   repository sublevels with respective class attributes.

TODO
}
\description{
Provides scaffold of a package repository as expected by R.
}
\section{Methods (by class)}{
\itemize{
\item \code{RappPackageRepositoryS3}: 

\item \code{missing}: 

\item \code{character}: 
}}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/repositr}
}
\seealso{
\code{\link[rapp2]{getExpandedRepositoryPaths-character-method}}
}

