% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getRepositoryPathByType}
\alias{getRepositoryPathByType}
\alias{getRepositoryPathByType,RappExpandedPackageRepositoryS3,character-method}
\alias{getRepositoryPathByType,RappExpandedPackageRepositoryS3,missing-method}
\alias{getRepositoryPathByType,RappPackageRepositoryMacBinaryS3,ANY-method}
\alias{getRepositoryPathByType,RappPackageRepositoryS3,character-method}
\alias{getRepositoryPathByType,RappPackageRepositoryS3,missing-method}
\alias{getRepositoryPathByType,RappPackageRepositorySourceS3,ANY-method}
\alias{getRepositoryPathByType,RappPackageRepositoryWinBinaryS3,ANY-method}
\alias{getRepositoryPathByType,character,character-method}
\alias{getRepositoryPathByType,character,missing-method}
\title{Select Repository Path By Type}
\usage{
getRepositoryPathByType(repos, type = c("win.binary", "source", "mac.binary"))


  \S4method{getRepositoryPathByType}{RappExpandedPackageRepositoryS3,character}(repos,
  type = c("win.binary", "source", "mac.binary"))


  \S4method{getRepositoryPathByType}{RappExpandedPackageRepositoryS3,missing}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{RappPackageRepositoryS3,character}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{character,character}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{character,missing}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{RappPackageRepositoryS3,missing}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{RappPackageRepositoryMacBinaryS3,ANY}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{RappPackageRepositoryWinBinaryS3,ANY}(repos,
  type = c("win.binary", "source", "mac.binary"))

\S4method{getRepositoryPathByType}{RappPackageRepositorySourceS3,ANY}(repos,
  type = c("win.binary", "source", "mac.binary"))
}
\arguments{
\item{repos}{\strong{Signature argument}.
Object containing repository information.}

\item{type}{\strong{Signature argument}.
Object containing repository type information.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{type}{\code{\link{character}}.}

\item{repos}{\code{\link{RappExpandedPackageRepositoryS3}}.}

\item{type}{\code{\link{missing}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{type}{\code{\link{character}}.}

\item{repos}{\code{\link{character}}.}

\item{type}{\code{\link{character}}.}

\item{repos}{\code{\link{character}}.}

\item{type}{\code{\link{missing}}.}

\item{repos}{\code{\link{RappPackageRepositoryS3}}.}

\item{type}{\code{\link{missing}}.}

\item{repos}{\code{\link{RappPackageRepositoryMacBinaryS3}}.}

\item{type}{\code{\link{ANY}}.}

\item{repos}{\code{\link{RappPackageRepositoryWinBinaryS3}}.}

\item{type}{\code{\link{ANY}}.}

\item{repos}{\code{\link{RappPackageRepositorySourceS3}}.}

\item{type}{\code{\link{ANY}}.}
}
\value{
\code{\link{character}}. Repository path

\code{\link{character}}. Repository path
}
\description{
Selects repository path based on tyype.
}
\section{Methods (by class)}{
\itemize{
\item \code{repos = RappExpandedPackageRepositoryS3,type = character}: 

\item \code{repos = RappExpandedPackageRepositoryS3,type = missing}: 

\item \code{repos = RappPackageRepositoryS3,type = character}: 

\item \code{repos = character,type = character}: 

\item \code{repos = character,type = missing}: 

\item \code{repos = RappPackageRepositoryS3,type = missing}: 

\item \code{repos = RappPackageRepositoryMacBinaryS3,type = ANY}: 

\item \code{repos = RappPackageRepositoryWinBinaryS3,type = ANY}: 

\item \code{repos = RappPackageRepositorySourceS3,type = ANY}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

