% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{normalizeRepositoryPath}
\alias{normalizeRepositoryPath}
\alias{normalizeRepositoryPath,character,character-method}
\alias{normalizeRepositoryPath,character,missing-method}
\title{Normalize Repository Path}
\usage{
normalizeRepositoryPath(path = ".", type = c("filesystem", "url"))

\S4method{normalizeRepositoryPath}{character,character}(path = ".",
  type = c("filesystem", "url"))

\S4method{normalizeRepositoryPath}{character,missing}(path = ".",
  type = c("filesystem", "url"))
}
\arguments{
\item{path}{\strong{Signature argument}.
Object containing path information.}

\item{type}{\strong{Signature argument}.
Object containing normalization type information.}

\item{path}{\code{\link{character}}.}

\item{type}{\code{\link{character}}.}

\item{path}{\code{\link{character}}.}

\item{type}{\code{\link{missing}}.}
}
\value{
\code{\link{character}}. Normalized path.

\code{\link{character}}. Normalized path.
}
\description{
Normalizes repository paths.
}
\section{Methods (by class)}{
\itemize{
\item \code{path = character,type = character}: 

\item \code{path = character,type = missing}: 
}}
\examples{
\dontrun{
  
  ## TODO: add example
  
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://www.rappster.de/rapp.core.repos}
}

